Add TS typing to `extendConnection()`
